Type part mutations

+++ src/store/part_setters.ts
@@ -4,14 +4,39 @@ import _ from '../util';
 import store from '../store';
 import LDParse from '../ld_parse';
 
-export default {
-	// opts: {partID, step, direction, partDistance=60, arrowOffset=0, arrowLength=60, arrowRotation=0}
+export interface PartMutationInterface {
+	displace({
+		partID, step, direction,
+		partDistance, arrowOffset, arrowLength, arrowRotation,
+	}: {
+		partID: number, step: LookupItem, direction?: Direction,
+		partDistance?: number, arrowOffset?: number, arrowLength?: number, arrowRotation?: number,
+	}): void;
+	moveToStep(
+		{partID, srcStep, destStep, doLayout}
+		: {partID: number, srcStep: LookupItem, destStep: LookupItem, doLayout?: boolean}
+	): void;
+	addToCallout(
+		{partID, step, callout, doLayout}
+		: {partID: number, step: LookupItem, callout: LookupItem, doLayout?: boolean}
+	): void;
+	removeFromCallout({partID, step}: {partID: number, step: LookupItem}): void;
+	delete(
+		{partID, step, doLayout}
+		: {partID: number, step: LookupItem, doLayout?: boolean}
+	): void;
+}
+
+export const PartMutations: PartMutationInterface = {
 	// If direction == null, remove displacement
 	displace({
 		partID, step, direction,
 		partDistance = 60, arrowOffset = 0, arrowLength = 60, arrowRotation = 0,
 	}) {
 		const stepItem = store.get.step(step);
+		if (stepItem == null || stepItem.csiID == null) {
+			return;
+		}
 		const displacementDistance = 60;
 		store.mutations.csi.resetSize({csi: stepItem.csiID});
 		partDistance = (partDistance == null) ? displacementDistance : partDistance;
@@ -45,10 +70,16 @@ export default {
 	},
 	moveToStep({partID, srcStep, destStep, doLayout = false}) {
 		const srcStepItem = store.get.step(srcStep);
+		if (srcStepItem == null || srcStepItem.csiID == null) {
+			return;
+		}
 		store.mutations.step.removePart({step: srcStepItem, partID});
 		store.mutations.csi.resetSize({csi: srcStepItem.csiID});
 
 		const destStepItem = store.get.step(destStep);
+		if (destStepItem == null || destStepItem.csiID == null) {
+			return;
+		}
 		store.mutations.step.addPart({step: destStepItem, partID});
 		store.mutations.csi.resetSize({csi: destStepItem.csiID});
 
@@ -62,21 +93,31 @@ export default {
 	addToCallout({partID, step, callout, doLayout = false}) {
 		const stepItem = store.get.step(step);
 		const calloutItem = store.get.callout(callout);
+		if (stepItem == null || calloutItem == null) {
+			return;
+		}
 		let destCalloutStep;
 		if (_.isEmpty(calloutItem.steps)) {
 			destCalloutStep = store.mutations.step.add({dest: calloutItem});
 		} else {
-			destCalloutStep = store.get.step(_.last(calloutItem.steps));
+			const lastStepId = _.last(calloutItem.steps) || -1;
+			destCalloutStep = store.get.step(lastStepId);
+		}
+		if (destCalloutStep == null) {
+			return;
 		}
 		destCalloutStep.model = _.cloneDeep(stepItem.model);
 		destCalloutStep.parts.push(partID);
 		store.mutations.csi.resetSize({csi: destCalloutStep.csiID});
 		if (doLayout) {
-			store.mutations.page.layout({page: step.parent});
+			store.mutations.page.layout({page: stepItem.parent});
 		}
 	},
 	removeFromCallout({partID, step}) {
 		const stepItem = store.get.step(step);
+		if (stepItem == null || stepItem.csiID == null) {
+			return;
+		}
 		_.deleteItem(stepItem.parts, partID);
 		store.mutations.csi.resetSize({csi: stepItem.csiID});
 		store.mutations.page.layout({page: store.get.pageForItem(stepItem)});
@@ -84,7 +125,13 @@ export default {
 	delete({partID, step, doLayout}) {
 		// Remove part from the step its in and from the model entirely
 		const partStep = store.get.step(step);
+		if (partStep == null) {
+			return;
+		}
 		const model = LDParse.model.get.abstractPart(partStep.model.filename);
+		if (model == null) {
+			return;
+		}
 		const part = LDParse.model.get.partFromID(partID, model.filename);
 		store.mutations.step.removePart({step, partID, doLayout});
 		store.mutations.inventoryPage.removePart({part, doLayout: doLayout});
